Add bus.js in the newly created directory utils to send value between two components. The sending part is in yearslider.js and receiving part is in imageExplorer.js.
Modified the yearslider.js to match the structure of imageExplorer.js (added a default constructor and initial state information). Added onChangeHandler to the slider to send the selected range of year to imageExplorer.js.

In the imageExplorer.js, added new function called componentDidMount() to receive value from yearslider and filter the image. Change the default showing from creating a new annotatedImages array to one newly created array in state(makes it possible to refresh the image).